requirements.txt
python-dotenv
langchain-community
langchain-chroma
langchain-experimental
langchain-openai
langchain
pgvector
psycopg2-binary
ipykernel
chromadb
typer[all]
pandas
pypdf
ollama